PLANET-7446: Expand images on click
- Apply lightbox to the grid layout
